+++ samples/elaborate/contact_manager.rb
@@ -52,7 +52,7 @@ class ContactManager
           }
         }
 
-        table { |table_proxy|
+        table(:multi) { |table_proxy|
           layout_data {
             horizontal_alignment :fill
             vertical_alignment :fill
@@ -83,7 +83,7 @@ class ContactManager
           }
           items bind(@contact_manager_presenter, :results),
           column_properties(:first_name, :last_name, :email)
-          on_mouse_up { |event|
+          on_mouse_down { |event|
             table_proxy.edit_table_item(event.table_item, event.column_index)
           }
         }