+++ CHANGELOG.md
@@ -5,7 +5,9 @@
 - Fix Glimmer::SWT::ShellProxy#pack_same_size for Linux
 - DSL keyword 'message_box'
 - Support table single selection databinding
+- Support table multi selection databinding
 - Support table cell editing databinding
+- Enhance table listener events with table_item and column_index methods
 
 ## 0.8.1